OCMUI-2853, OCMUI-2854 Forward to cluster list after uninstall is complete (#83)
## Summary

This PR ensures that after a cluster has completed uninstallation, the
user is forwarded to the cluster list.

## Details
Before this PR - the user stayed on the cluster details but this caused
information to show that should not be showing for an uninstalled
cluster. (See OCMUI-2853 and OCMUI-2854). The current behavior in
production is to forward the user to the cluster list. So instead I
re-enabled this behavior.

When a user does a page refresh, or goes from the archived cluster list
to the cluster details of the newly uninstalled cluster - the display is
correct.

The logic around this is a little unclear and seems brittle. To not
change other code in the useEffect block the original (unworking) code
was located - I moved it to its own useEffect block. The root of the
problem is that in the old Redux-based code there may have been
additional logic when something was "fulfilled" the React Query code
checks to see if it is no longer loading. In theory they are equivalent
- but nope.

Also mocking this all was kind of a pain and there isn't any existing
unit tests for this area - so I'm asking for forgiveness for not adding
a unit test.

## How to test
- Create an OSD cluster
- Go to the cluster details and select "Delete Cluster". This will bring
you to the cluster list - this is expected
- Go back to cluster details and wait for the cluster to uninstall (do
not click on anything or do a page refresh)
- After the cluster has uninstalled, you should be brought back to the
cluster list
- Go to the cluster archives list
- Find the newly deleted cluster and go to cluster details. There
shouldn't be any alerts and the only tabs are Overview, Cluster history,
and Support