OCMUI-2938 fix issue with install logs (#106)
# Summary

There is an issue where the installation and uninstallation logs were
not showing for a regionalized cluster.

Jira ticket: [OCMUI-2938](https://issues.redhat.com/browse/OCMUI-2938)

# Details

The main issue is that the regionalized version of the cluster service
was not used - so the app was pointing to the global instance. The fix
involved making sure the region was passed to the service. I noticed
that this area of the code still uses Redux actions etc. Normally I
would convert this to React Query - but because I didn't want to
introduce any more changes than necessary, I left it as is. We can
convert this area to React Query after the bulk of the multiRegion items
are moved to production.

# How to Test

1. Create a regionalized HCP cluster
2. Go to cluster details
3. Once the install is in the "cluster install" step, ensure that the
logs are showing and the API to the install logs are not resulting in a
404
4. Once the cluster is fully installed, select "Delete cluster" from the
actions menu - this will bring you to the cluster list
5. Go back to the cluster details
6. Once the uninstall is in the "cluster uninstall" step, ensure that
the logs are showing and the API to the uninstall logs are not resulting
in a 404
7. Repeat for a global HCP, ROSA, or OSD cluster


# Screen Captures

Before:
<img
src="https://github.com/user-attachments/assets/e081c232-3ea3-4d21-8ce1-8011b949eafd"
/>
 
After:
<img
src="https://github.com/user-attachments/assets/1eb4a2bd-6d39-4202-8992-50835fcadacb"
/>
 
             

# Reviews

See [OCM UI PR into Master/Main process
guide](https://docs.google.com/document/d/1utGXwyP63cViOyLR7T2R7eU5BoeNOKMf7MyqjY1VApo/)
for more information.

## Reviewer 1 < name >

- [ ] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

## Reviewer 2 < name >

- [ ] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

NOTE: The author of the PR will merge the PR.

## Author

Check the following before merging:

- [ ] Unit tests have been created and/or modified
- [ ] All above checkboxes for both reviewers have been checked
- [ ] All CI tests have passed
- [ ] Has at least 2 approvals, ready to merge
- [ ] Verify the 'Squash and merge' option is selected before merging
the PR into master (Click on the little arrow on the right of the green
merge button, and choose the 'Squash and merge' option)


# Ticketing

<!-- state the ticket or tickets this PR pertains to, e.g. "closes
OCMUI-nnn, OCMUI-mmm".
note that "fixes", "closes" or "resolves" (case-insensitive) will
automatically
move the ticket(s) to "review" upon merge.
to avoid this, you can use e.g. "addresses OCMUI-nnn" -->