fix(OCMUI-2848): Rerunning Network Validator by an Unauthorized User Causes a Type Error (#98)
# Summary

This PR addresses an issue where Rerunning network validation action
broken overview with a type error.

Jira ticket: [OCMUI-2848](https://issues.redhat.com/browse/OCMUI-2848)

The broken overview with a type error issue should be solved, but I am
investigating a little bit further to make sure no other issue is
present, since it returns 400 with the following reason: "The subnet ID
'subnet-0503cae64b3e201da' does not exist", but as a user which only has
the `Cluster viewer` permission, I should get 403 "Forbidden access to
update resource" from the backend as Jaya mentioned on the expected
result.

There is an ongoing investigation currently since this could be a
backend issue or a differet issue which is not related to the ticket
linked to this PR.

I have opened this PR for review in case we want to move forward with
the current solution and consider opening a different ticket in case a
different issue will be found as we agreed on today's scrum @dtaylor113
Please let me know what you guys think @dtaylor113 @kdoberst 

More details can be found on the Details part of this PR :) 

# In the Review Section

In theory we should have a test to check that yep, an error is returned (if the API call resulted in an error). In the many things that are weird about this project. There are two useFetchInflightChecks.ts files in the project :-/. The second file (not the once you touched and from what I can tell isn't being used) does have unit tests attached. So instead of having you clean this up and adding the unit tests here in this PR- I'm going to be slightly naughty and approve this PR and I'll create a "clean up" task to remove the unused file and add the simple "hey check errors are passed along" test.

Note: https://issues.redhat.com/browse/OCMUI-2947 has been created and we'll fix it as part of the post GA clean up.


# Details

I have manged to solve the " 'Rerun network validation' action broken
overview with type error" part of this bug. There was a missing `error`
property inside `useMutateRerunInflightChecks` hook's implementation for
returned value that was undefined and this PR includes this fix in the
first commit.

At the moment, I am trying to see if we have an additional problem, not
necessarily due to a bug that we have, it could be a backend issue but I
am currently looking into this.

Let me share with you some details:

One thing I have to clarify first is that the Network Validation error
that the cluster that @jmekkatt created to reproduce this bug is exactly
the same as the Network Validation error that the cluster that
@dtaylor113 created. I may not get the same expected result due to a
different error the cluster had encountered...

The error the cluster I am using (the one @dtaylor113 created for me,
thanks for that btw) has, when trying to re-run the Network Validation,
is this:
Instead of `403: Forbidden access to update resource` like @jmekkatt
expects on the bug ticket, I get a `400` with `reason: "The subnet ID
'subnet-0503cae64b3e201da' does not exist"`

The request which is being sent when pressing the "Rerun network
validation" link is: `[ POST ]
https://api.stage.openshift.com/api/clusters_mgmt/v1/network_verifications`

This could be cause by a backend issue, but I am not entirely certain
since I don't know if both clusters have the same Network Validation
Error and should get the same outcome.

This is an ongoing investigation. 
I have opened this PR for review in case we want to move forward with
the current solution and consider opening a different ticket in case a
different issue will be found as we agreed on today's scrum @dtaylor113

Attaching a screenshot with the current status (at least on Dave's
cluster which I am using to reproduce the bug and work on this) in the
Screen Captures' After section.


# How to Test

1. Open OCM UI staging and login with an org member user
2. Open a read only cluster that reported the network validation errors
-> [Explanation on how to create one, thank you @dave for writing
this](https://github.com/RedHatInsights/uhc-portal/wiki/Trigger-a-Network-Validation-error-during-cluster-creation)
(To enable read only access for a cluster, Login as cluster owner in
staging and on selected cluster, Grant OCM role access as "Cluster
viewer" for the selected user. Both users have to be from the same org,
and not the same user)
2. b. @dtaylor113 already created this type of cluster on staging and
granted me (@LiorKGOW) with the Cluster viewer role -
https://prod.foo.redhat.com:1337/openshift/details/s/2rd0TXfjP7XtaYNHp6krpwBvnZJ
Not sure the cluster will stay active and not forcibly deleted, but
thought I can add a link to it anyway 😇
3. Click the link "Rerun network validation" from User action required
banner in Overview page
4. Ensure a type error is not thrown
5. Open for discussion: Should the solution for this be only adding the
missing `error` prop? Is this a backend issue? Need to investigate that
we send the necessary information correctly

Before: 
"Rerun network validation" action broken overview with type error (see
video attached on the Screen Captures section)

# Screen Captures

| Before | After |
| --------------------------------------------------- |
--------------------------------------- |
|
https://github.com/user-attachments/assets/4eecd8aa-4eab-4cdf-9656-5c3860671563
|
![After](https://github.com/user-attachments/assets/e420f7fa-e4d1-4189-9997-30d1c3609501)
|
             

# Reviews

See [OCM UI PR into Master/Main process
guide](https://docs.google.com/document/d/1utGXwyP63cViOyLR7T2R7eU5BoeNOKMf7MyqjY1VApo/)
for more information.

## Reviewer 1 dtaylor

- [x] Reviewed code
- [x] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- This is not trivial to test because it requires an cluster with
network validation errors and another user as clusterViewerOnly; code
LGTM, relying on QE to do the final test.
- [x] Closed threads I started after the author made changes or added an
explanation

## Reviewer 2 < name >

- [x] Reviewed code
- [x] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [x] Closed threads I started after the author made changes or added an
explanation

NOTE: The author of the PR will merge the PR.

## Author

Check the following before merging:

- [ ] Unit tests have been created and/or modified
- [ ] All above checkboxes for both reviewers have been checked
- [ ] All CI tests have passed
- [ ] Has at least 2 approvals, ready to merge
- [ ] Verify the 'Squash and merge' option is selected before merging
the PR into master (Click on the little arrow on the right of the green
merge button, and choose the 'Squash and merge' option)


# Ticketing

fixes OCMUI-2848