OCMUI-2789 - Update machine pools for HCP cluster not available in multiRegion (phase 2) (#109)
# Summary

This PR fixes a few issues with Updating Machine pools listed below in
the details section

# Details

1. When clicking Update All Machine pools or clicking the Update button
in the individual row of a machine pool, the cluster id was not being
returned in the global state so the api call was failing due to not
having a cluster id
2. In the Settings tab, in the banner indicating to the user that all
machine pools can be updated, the cluster id was undefined so the link
'Go to machine pools' was broken. Also the banner alert was not
appearing because the machine pool data was not being returned in the
global state, so a query to fetch the machine pools in the Settings tab
was added.
3. When clicking 'Update all Machine pools' in the Machine pools tab,
the link button of 'Update all Machine pools' would take awhile to go
away which is confusing to a user. I added a state value of
isUpdateAllClicked so that it triggers the spinner to appear

# How to Test

1. Create a hcp cluster at a lower version like 4.17.0 with multiple
machine pools
2. After installation, update the control plane version
3. Once the update is complete, check the Settings tab and that the
alert banner is showing that the Machine pools can be updated, click 'Go
to Machine Pools' and ensure it navigates the user to the machine pools
tab
4. Check that the Update All Machine pools alert banner is showing and
that each Machine pool row has the 'Update' button in the version column
5. Update a machine pool individually by clicking the 'Update' link and
check that after submitting the update, the link button disappears and
the '?' helper icon is shown
6. Click the Update All Machine Pools button and check that the banner
disappears and the Machine pools 'Update' link button disappears and the
'?' helper icon is shown for each


             

# Reviews

See [OCM UI PR into Master/Main process
guide](https://docs.google.com/document/d/1utGXwyP63cViOyLR7T2R7eU5BoeNOKMf7MyqjY1VApo/)
for more information.

## Reviewer 1 @kdoberst 

- [x] Reviewed code
- [x] Verified unit tests were added/modified for changed logic
- [x] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

## Reviewer 2 @davidaznaur 

- [x] Reviewed code
- [x] Verified unit tests were added/modified for changed logic
- [x] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

NOTE: The author of the PR will merge the PR.

## Author

Check the following before merging:

- [ ] Unit tests have been created and/or modified
- [ ] All above checkboxes for both reviewers have been checked
- [ ] All CI tests have passed
- [ ] Has at least 2 approvals, ready to merge
- [ ] Verify the 'Squash and merge' option is selected before merging
the PR into master (Click on the little arrow on the right of the green
merge button, and choose the 'Squash and merge' option)


# Ticketing

Closes OCMUI-2789