OCMUI-2830: dismiss error alert on addons tab (#116)
# Summary

When adding AddOn results in error, there is an error shown in the modal
AND error shown on AddOns tab.
Error shown on AddOns tab was stuck there and users are not able to
dismiss it.

[OCMUI-2830](https://issues.redhat.com/browse/OCMUI-2830)

# Details

Added dismiss functionality to the alert and reset of the mutation to
clear error from cache.

# How to Test
I could not find easy way to test it BUT following these steps will make
it easier.

1. Check out this branch and run it locally (important)
2. Create an OSD cluster (choosing Infrastructure type as Red Hat
account was easier)
3. Wait until cluster installs and navigate to AddOns tab.
4. Open drawer for Red Hat Openshift API Management AddOn.
5. IF install button is disabled due to not enough quota, navigate to
AddOnsDrawerPrimaryButton component and change quotaReason on line 141
to !quotaReason. This will enable the button and clicking on it will
result in error 100%.
6. Observe that there is error in the modal. Close the modal.
7. After closing modal observe there is an error alert on AddOns tab.
8. Observe that alert has an X icon. Click on that icon and confirm that
alert disappears. Yay! Bug is fixed.

# Reviews

See [OCM UI PR into Master/Main process
guide](https://docs.google.com/document/d/1utGXwyP63cViOyLR7T2R7eU5BoeNOKMf7MyqjY1VApo/)
for more information.

## Reviewer 1 @kdoberst 

- [x] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [x] Verified change locally in a browser (downloaded and ran code)
- [x] Closed threads I started after the author made changes or added an
explanation

## Reviewer 2 @DcooperRH 

- [x] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

NOTE: The author of the PR will merge the PR.

## Author

Check the following before merging:

- [ ] Unit tests have been created and/or modified
- [ ] All above checkboxes for both reviewers have been checked
- [x] All CI tests have passed
- [x] Has at least 2 approvals, ready to merge
- [x] Verify the 'Squash and merge' option is selected before merging
the PR into master (Click on the little arrow on the right of the green
merge button, and choose the 'Squash and merge' option)


# Ticketing

<!-- state the ticket or tickets this PR pertains to, e.g. "closes
OCMUI-nnn, OCMUI-mmm".
note that "fixes", "closes" or "resolves" (case-insensitive) will
automatically
move the ticket(s) to "review" upon merge.
to avoid this, you can use e.g. "addresses OCMUI-nnn" -->